add function and do it as a separate file
